// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/paymentResponseStatus.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type PaymentResponseStatus = typeof PaymentResponseStatus[keyof typeof PaymentResponseStatus];


export const PaymentResponseStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
} as const;